# src/public/swagger/Auth.yaml
openapi: 3.0.0
info:
  title: School Management API
  version: 1.0.0
  description: API documentation for the School Management System, including authentication and core entities like Users, Kelas, Siswa, and Guru.
...
paths:
...
  /api/v1/kelas/{id_kelas}:
    get:
      summary: Get a kelas by ID
      tags: [Kelas]
      parameters:
        - in: path
          name: id_kelas
          schema:
            type: integer
          required: true
          description: Numeric ID of the kelas to retrieve
      responses:
        200:
          description: Class found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KelasResponse'
        404:
          description: Class not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    put:
      summary: Update a kelas by ID
      tags: [Kelas]
      parameters:
        - in: path
          name: id_kelas
          schema:
            type: integer
          required: true
          description: Numeric ID of the kelas to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Kelas'
      responses:
        200:
          description: Class updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KelasResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        404:
          description: Class not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    delete:
      summary: Delete a kelas by ID
      tags: [Kelas]
      parameters:
        - in: path
          name: id_kelas
          schema:
            type: integer
          required: true
          description: Numeric ID of the kelas to delete
      responses:
        200:
          description: Class deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
        404:
          description: Class not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
...
  /api/v1/jurusan/{id_jurusan}:
    get:
      summary: Get a jurusan by ID
      tags: [Jurusan]
      parameters:
        - in: path
          name: id_jurusan
          schema:
            type: integer
          required: true
          description: Numeric ID of the jurusan to retrieve
      responses:
        200:
          description: Major found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JurusanResponse'
        404:
          description: Major not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    put:
      summary: Update a jurusan by ID
      tags: [Jurusan]
      parameters:
        - in: path
          name: id_jurusan
          schema:
            type: integer
          required: true
          description: Numeric ID of the jurusan to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Jurusan'
      responses:
        200:
          description: Major updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JurusanResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        404:
          description: Major not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    delete:
      summary: Delete a jurusan by ID
      tags: [Jurusan]
      parameters:
        - in: path
          name: id_jurusan
          schema:
            type: integer
          required: true
          description: Numeric ID of the jurusan to delete
      responses:
        200:
          description: Major deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
        404:
          description: Major not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
...
  /api/v1/guru/{id_guru}:
    get:
      summary: Get a guru by ID
      tags: [Guru]
      parameters:
        - in: path
          name: id_guru
          schema:
            type: integer
          required: true
          description: Numeric ID of the guru to retrieve
      responses:
        200:
          description: Teacher found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuruResponse'
        404:
          description: Teacher not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    put:
      summary: Update a guru by ID
      tags: [Guru]
      parameters:
        - in: path
          name: id_guru
          schema:
            type: integer
          required: true
          description: Numeric ID of the guru to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Guru'
      responses:
        200:
          description: Teacher updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuruResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        404:
          description: Teacher not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    delete:
      summary: Delete a guru by ID
      tags: [Guru]
      parameters:
        - in: path
          name: id_guru
          schema:
            type: integer
          required: true
          description: Numeric ID of the guru to delete
      responses:
        200:
          description: Teacher deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
        404:
          description: Teacher not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StandardSuccessResponse:
      type: object
      properties:
        status:
          type: boolean
          description: Indicates whether the request was successful
          example: true
        message:
          type: string
          description: A descriptive message
          example: Operation successful
      required:
        - status
        - message
    StandardErrorResponse:
      type: object
      properties:
        status:
          type: boolean
          description: Indicates whether the request was successful
          example: false
        message:
          type: string
          description: A descriptive error message
          example: An error occurred
      required:
        - status
        - message

    # --- Core Entity Schemas (based on Prisma) ---
...
    Kelas:
      type: object
      required:
        - nama_kelas
        - kelas_romawi
      properties:
        id_kelas:
          type: integer
          readOnly: true
          description: Unique identifier for the kelas.
          example: 1
        nama_kelas:
          type: string
          description: Name of the kelas (e.g., "10 IPA 1").
          example: "10 IPA 1"
        kelas_romawi:
          type: string
          description: Roman numeral representation of the kelas level (e.g., "X", "XI", "XII").
          example: "X"
      example:
        nama_kelas: "10 IPA 1"
        kelas_romawi: "X"
    KelasResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Class fetched successfully
        data:
          $ref: '#/components/schemas/Kelas'
...
    Jurusan:
      type: object
      required:
        - nama_jurusan
        - deskripsi
      properties:
        id_jurusan:
          type: integer
          readOnly: true
          description: Unique identifier for the jurusan.
          example: 1
        nama_jurusan:
          type: string
          description: Name of the jurusan (e.g., "IPA", "IPS", "Bahasa").
          example: "IPA"
        deskripsi:
          type: string
          description: Description of the jurusan.
          example: "Ilmu Pengetahuan Alam"
      example:
        nama_jurusan: "IPS"
        deskripsi: "Ilmu Pengetahuan Sosial"
    JurusanResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Major fetched successfully
        data:
          $ref: '#/components/schemas/Jurusan'
...
    Guru:
      type: object
      required:
        - nip
        - nama_guru
        - id_mapel
      properties:
        id_guru:
          type: integer
          readOnly: true
          description: Unique identifier for the guru.
          example: 1
        nip:
          type: string
          description: National Educator Identification Number (Nomor Induk Pendidik) for the guru.
          example: "198001012005011001"
        nama_guru:
          type: string
          description: Full name of the guru.
          example: "Dra. Siti Aminah"
        id_mapel:
          type: integer
          description: ID of the mata-pelajaran this guru teaches.
          example: 101
      example:
        nip: "198001012005011001"
        nama_guru: "Dra. Siti Aminah"
        id_mapel: 1
    GuruResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Teacher fetched successfully
        data:
          $ref: '#/components/schemas/Guru'
...